241105 양혁수 이재서 pr 저장

+++ dr/src/main/java/com/dr/dto/board/FreeBoardListDTO.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 @ToString
 @NoArgsConstructor
 public class FreeBoardListDTO {
-    private Long boardNumber;
+    private int boardNumber;
     private String boardTitle;
     private String userNickName;
     private int goodCount;
